+ initial implementation for unassigned load
git-svn-id: http://agilefant.googlecode.com/svn/trunk@2573 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/transfer/UnassignedLoadTO.java
@@ -7,7 +7,7 @@ public class UnassignedLoadTO {
     public ExactEstimate effortLeft;
     public int iterationId;
     public Iteration iteration;
-    public short availabilitySum;
+    public int availabilitySum;
     public short availability;
     
     public UnassignedLoadTO(ExactEstimate effortLeft, int iterationId, short userAvailability) {
